[REFACTOR] application.timeline 패키지에 타임라인 조회 기능 이동 #107
