+++ src/main/java/com/sports/server/query/application/timeline/ScoreRecordQueryService.java
@@ -1,4 +1,4 @@
-package com.sports.server.query.application;
+package com.sports.server.query.application.timeline;
 
 import com.sports.server.query.dto.mapper.RecordMapper;
 import com.sports.server.query.dto.response.RecordResponse;
